icam-abstract-db/backend/requirements.txt
beautifulsoup4==4.12.3
Flask-Cors==4.0.0
Flask==3.0.2
requests==2.31.0
sentence-transformers==2.7.0
redis==5.0.4
elasticsearch==8.12.1
faiss-cpu==1.80